Move out more nitextureproperty code to specialised class away from generic TextureSlotManager
